[fix] 引数から指定できるように修正

+++ buffer/cpu/rs/src/bindings/jni/runtime.rs
@@ -1,6 +1,6 @@
 use jni::{JNIEnv};
 use jni::objects::{JClass};
-use jni::sys::{jint, jlong};
+use jni::sys::jlong;
 
 use crate::ops::runtime;
 use crate::runtime::Runtime;
@@ -9,7 +9,7 @@ use crate::runtime::Runtime;
 pub extern "system" fn Java_com_wsr_knist_cpu_JRuntime_allocate(
     _: JNIEnv,
     _class: JClass,
-    pool_size: jint,
+    pool_size: jlong,
 ) -> jlong {
     let runtime = runtime::allocate(pool_size as u64);
     Box::into_raw(Box::new(runtime)) as jlong
