+++ buffer/cpu/rs/src/bindings/native/headers/runtime_native.h
@@ -3,7 +3,7 @@
 
 typedef struct Runtime Runtime;
 
-Runtime* com_wsr_cpu_runtime_alloc(int pool_size);
+Runtime* com_wsr_cpu_runtime_alloc(long pool_size);
 void com_wsr_cpu_runtime_release(Runtime* runtime);
 
 #endif
